+++ src/DnaPSSM.cpp
@@ -445,7 +445,7 @@ void DnaPSSM::integrate_energy_logspat(const string &target, float &energy, vect
 					break;
 				}
 				if(DnaLookupTables::NEUTRAL_CHAR[(unsigned char)*j]) {
-					logp += c_log_quarter;
+					logp += p->get_avg_log_prob();
 				} else {
 					int code = DnaLookupTables::COMPLEMENT_ENCODE[(unsigned char)*j];
 					if(code >= 0) {
@@ -517,7 +517,7 @@ void DnaPSSM::integrate_energy_max_logspat(const string &target, float &energy, 
 					break;
 				}
 				if(DnaLookupTables::NEUTRAL_CHAR[(unsigned char)*j]) {
-					logp_rev += c_log_quarter;
+					logp_rev += p->get_avg_log_prob();
 				} else {
 					int code = DnaLookupTables::COMPLEMENT_ENCODE[(unsigned char)*j];
 					if(code >= 0) {